// src/theme/theme.types.ts
export type Theme = {
  ui: {
    bg: string
    sidebar: string
    activityBar: string
    statusBar: string
    panel: string
    text: string
    tabActive: string
    tabBorder: string
    border: string
  }

  editor: {
    bg: string
    text: string
    comment: string
    keyword: string
    string: string
    function: string
    variable: string
    number: string
    type: string
    lineHighlight: string
  }
}